config-1-transfer-code-from-csd-service-repository-to-this-repository: Added poetry toml and fixed imports and file pathes BUT still need fixture for the tests because of huggingface type problems
