+++ pyproject.toml
@@ -1,0 +1,34 @@
+[tool.poetry]
+name = "chess-scoresheet-digitization-training"
+version = "0.0.1"
+description = ""
+authors = ["Benjamin Kostka mail@ben-kostka.de"]
+readme = "README.md"
+packages = [{include = "src"}]
+
+[tool.poetry.dependencies]
+python = "^3.12"
+pytest = "*"
+fastapi = "*"
+uvicorn = "*"
+python-multipart = "*"
+aiofiles = "*"
+mlflow = "*"
+numpy = "*"
+opencv-python = "*"
+pandas = "*"
+starlette = "*"
+dagshub = "*"
+datasets = "*"
+transformers = "*"
+seaborn = "*"
+ibm_watson_machine_learning = "*"
+pre-commit = "*"
+matplotlib = "*"
+joblib = "*"
+pyyaml = "*"
+torch = "*"
+
+[build-system]
+requires = ["poetry-core>=2.0.0,<3.0.0"]
+build-backend = "poetry.core.masonry.api"
