--- requirements.txt
@@ -1,26 +1,0 @@
-fastapi
-uvicorn
-datasets
-python-multipart
-aiofiles
-click
-fastapi
-mlflow
-mlflow_skinny
-numpy
-opencv-python
-pandas
-Pillow
-scikit_learn
-starlette
-transformers
-dagshub
-seaborn
-ibm_watson_machine_learning
-pytest
-pytest-dependency
-pre-commit
-matplotlib
-joblib
-pyyaml
-torch
