--- src/tests/test_chess-scoresheet-digitalization-training.py
@@ -1,14 +1,0 @@
-#!/usr/bin/env python
-
-# Example Tests
-
-import pytest
-
-
-def capital_case(x):
-    return x.capitalize()
-
-
-def test_capital_case():
-    assert capital_case("semaphore") == "Semaphore"
-
